data_storage_service/requirements.txt
grpcio
grpcio-tools
protobuf
aiofiles
scipy
matplotlib
redis
pillow


